chapter_007/src/main/java/ru/job4j/lambda/FunctionCalculator.java
package ru.job4j.lambda;

import java.util.ArrayList;
import java.util.List;
import java.util.function.Function;

/**
 * FunctionCalculator
 *
 * @author Ksenya Kaysheva (murrcha@me.com)
 * @version $Id$
 * @since 0.1
 */
public class FunctionCalculator {

    /**
     * Method diapason - return list values of result func
     * @param start diapason
     * @param end diapason
     * @param func function
     * @return list
     */
    public List<Double> diapason(int start, int end, Function<Double, Double> func) {
        List<Double> result = new ArrayList<>();
        for (int index = start; index <= end; index++) {
            result.add(func.apply((double) index));
        }
        return result;
    }
}
